pyproject.toml
[project]
name = "damsso"
version = "0.1.0"
description = "DAMSSO (Django Allauth Multitenant SSO) - Django-allauth extension providing dynamic multi-tenant SSO support with OIDC and SAML"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "William Hayes", email = "william.s.hayes@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "django>=4.2",
    "django-allauth>=0.57.0",
    "python3-saml>=1.15.0",
    "authlib>=1.3.0",
    "cryptography>=41.0.0",
    "django-rls>=0.4.1",
    "uuid-utils>=0.9.0",
]
keywords = ["django", "allauth", "sso", "saml", "oidc", "multi-tenant"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Security",
]

[project.urls]
Homepage = "https://github.com/wshayes/damsso"
Documentation = "https://wshayes.github.io/damsso/"
Repository = "https://github.com/wshayes/damsso"
Issues = "https://github.com/wshayes/damsso/issues"

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.1.0",
    "responses>=0.23.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "pymdown-extensions>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/damsso"]
omit = ["*/tests/*", "*/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
